+++ website/src/main/scala/io/frontroute/site/examples/CodeExample.scala
@@ -2,7 +2,12 @@ package io.frontroute.site.examples
 
 import com.raquo.laminar.api.L._
 
-abstract class CodeExample(val id: String, val title: String, val description: String)(
+abstract class CodeExample(
+  val id: String,
+  val title: String,
+  val description: String,
+  val links: Seq[String]
+)(
   _code: sourcecode.Text[() => Element]
 ) {