+++ website/src/main/scala/io/frontroute/site/examples/ex_params/ParamsExample.scala
@@ -9,7 +9,15 @@ object ParamsExample
     extends CodeExample(
       id = "query-parameters",
       title = "Query parameters",
-      description = FileAsString("description.md")
+      description = FileAsString("description.md"),
+      links = Seq(
+        "/",
+        "/movie?id=2356777",
+        "/movie?id=0306414",
+        "/movies/search?director=cameron&year=1991",
+        "/movies/search?director=cameron",
+        "/movies/search?year=1991"
+      )
     )(() => {
       import com.raquo.laminar.api.L._
       import io.frontroute._
@@ -44,46 +52,6 @@ object ParamsExample
               div(unmatched.mkString("/", "/", ""))
             )
           }
-        ),
-        div(
-          cls := "bg-blue-900 -mx-4 -mb-4 p-2",
-          div(
-            cls := "font-semibold text-2xl text-blue-200",
-            "Navigation"
-          ),
-          div(
-            cls := "flex flex-col p-2",
-            a(
-              cls  := "text-blue-300 hover:text-blue-100",
-              href := "/",
-              "➜ /"
-            ),
-            a(
-              cls  := "text-blue-300 hover:text-blue-100",
-              href := "/movie?id=0306414",
-              "➜ /movie?id=0306414"
-            ),
-            a(
-              cls  := "text-blue-300 hover:text-blue-100",
-              href := "/movie?id=0103064",
-              "➜ /movie?id=0103064"
-            ),
-            a(
-              cls  := "text-blue-300 hover:text-blue-100",
-              href := "/movies/search?director=cameron&year=1991",
-              "➜ /movies/search?director=cameron&year=1991"
-            ),
-            a(
-              cls  := "text-blue-300 hover:text-blue-100",
-              href := "/movies/search?director=cameron",
-              "➜ /movies/search?director=cameron"
-            ),
-            a(
-              cls  := "text-blue-300 hover:text-blue-100",
-              href := "/movies/search?year=1991",
-              "➜ /movies/search?year=1991"
-            )
-          )
         )
       )
     })