+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * memoize is gone (event streams in routes interfere with `noneMatched`)
 * `when` --> `whenTrue` (conflict with laminext)
 * `mapOption` -> `optionMap`
-* 
+* `state` directive is gone
+
 ### 0.15.2
 
 Airstream `v0.14.2`.